backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
qiskit>=1.0.0
qiskit-aer>=0.13.0
openai>=1.0.0
numpy>=1.26.0
brainflow>=5.10.0
python-dotenv>=1.0.0
